Added session check on page load

+++ src/assets/js/routerCond.js
@@ -8,6 +8,9 @@ export default function(router, http) {
     console.error(endAt - Date.now())
     timeoutSet = true
     setTimeout(function() {
+      if (!timeoutSet) {
+        return
+      }
       alert('Session expired')
       timeoutSet = false
       storage.destroy()
@@ -15,10 +18,46 @@ export default function(router, http) {
     }, endAt - Date.now())
   }
 
+  const checkSession = function(title, next, first) {
+    // check for session
+    http.post('/checkSession').then(res => {
+      const setAt = res.data.setAt
+      const endAt = res.data.endAt
+
+      const isSet = setAt || endAt
+
+      // if i need auth but is not set
+      if (!isSet) {
+        // destroy session if on load
+        if (typeof first === 'boolean' && first === true) {
+          timeoutSet = false
+          storage.destroy()
+        }
+        router.push('login')
+        return
+      }
+
+      storage.set({ setAt: setAt, endAt: endAt })
+      expireTimer(endAt)
+
+      // must have id
+      document.title = title
+      next()
+    }).catch(e => {
+      next(false)
+    })
+
+  }
+
   // router
   router.beforeEach((to, from, next) => {
     // set title
     const title = to.meta.title || to.name
+
+    // make the request on refresh / page load
+    window.onload = function() {
+      checkSession(title, next, true)
+    }
 
     // check session, if component to needs authorization
     const cAuth = typeof to.meta.auth === 'undefined' ? true : to.meta.auth
@@ -58,29 +97,7 @@ export default function(router, http) {
     }
 
     // below here if all cAuth = true
-    // check for session
-    http.post('/checkSession').then(res => {
-      const setAt = res.data.setAt
-      const endAt = res.data.endAt
-
-      const isSet = setAt || endAt
-
-      // if i need auth but is not set
-      if (!isSet) {
-        router.push('login')
-        return
-      }
-
-      storage.set({ setAt: setAt, endAt: endAt })
-      expireTimer(endAt)
-
-      // must have id
-      document.title = title
-      next()
-    }).catch(e => {
-      next(false)
-    })
-
+    checkSession(title, next)
   })
-
+  
 }
